+++ jest.config.ts
@@ -1,0 +1,8 @@
+export default {
+  testEnvironment: 'jsdom',
+  transform: {
+    "^.+\\.(t|j)sx?$": "@swc/jest",
+  },
+  collectCoverage: true,
+  moduleFileExtensions: ['js', 'ts', 'node'],
+}